requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
asyncpg==0.28.0
pydantic==2.4.2
playwright==1.39.0
beautifulsoup4==4.12.2
requests==2.31.0
tweepy==4.14.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==3.0.1
celery==5.3.4
redis==5.0.1 